+++ inst/docker/rocker_scripts/install_extra_pkgs.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e
-cat pkgs-cran | grep -v "#" | xargs install2.r --error --skipinstalled
+cat pkgs-cran | grep -v "#" | xargs install2.r --error --skipinstalled --ncpus -1 --deps TRUE
 cat pkgs-github | grep -v "#" | xargs installGithub.r --deps TRUE